IA/ms-sdk-m365-agent-copilot-studio/src/requirements.txt
python-dotenv
microsoft-agents-hosting-aiohttp
microsoft-agents-authentication-msal
openai
httpx
tenacity
